move all use cases for Tasks into separate controllers
