+++ src/pages/Tasks.tsx
@@ -3,7 +3,8 @@ import styled from 'styled-components';
 import backbuttonsvg from '../assets/arrow-left-circle.svg';
 import addbuttonsvg from "../assets/plus-button.svg";
 import { Route, Routes, useNavigate, useParams } from 'react-router';
-import { getTasksAndTagsOfProperty, getTasksOfProperty, toggleTaskStatus } from '../controllers/TaskController';
+import { getTasksAndTagsOfProperty, getTasksOfProperty } from '../controllers/GetTaskController';
+import { toggleTaskStatus } from "../controllers/ToggleTaskStatusController";
 import AddEditTask from './AddEditTask';
 import Popup from '../components/Popup';
 import { Database } from "../supabase/supabase";
